+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a footer containing legal information to the login wizard
 - Added a notification to the timeline when a user's power level changes
 - Introduce new UI component for dropdown select
+- Add a settings panel for configuring power levels
 
 ### Changed